+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [itachallenge-user-3.1.3-RELEASE] - 2025-12-12
+
+### Changed
+- Refactored the logic for adding favorites to integrate it into the FavoriteService structure (without controller) (Taiga US [#904], Taiga Task [#917], PR [#1052])
+- Refactored the logic for deleting favorites to integrate it into the FavoriteService structure (with controller) (Taiga US [#904], Taiga Task [#918], PR [#1053])
+- Moved add favorite controller logic from UserController to FavoriteController (Taiga US [#904], Taiga Task [#953], PR [#1058])
+
 ### [itachallenge-user-3.1.2-RELEASE] - 2025-12-04
 
 ### Changed
@@ -41,9 +48,6 @@ Supported actions: SAVE, GIVE_UP, SUBMIT. (Taiga [#871], PR [#1043])
 
 ## [Unreleased]
 
-### Changed
-- [Changed] Refactored the logic for adding favorites to integrate it into the FavoriteService structure (Taiga [#917], PR [#1052])
-- [Changed] Refactored the logic for adding favorites to integrate it into the FavoriteService structure (Taiga [#918], PR [#1053])
 
 ### Chore/Internal
 - Introduced `SolutionAction` enum (internal change)